refactor: Separate rows render logic and implement RewGroup component
