+++ package.json
@@ -75,8 +75,6 @@
     "html-webpack-plugin": "^5.5.0",
     "jest": "^29.4.2",
     "mini-css-extract-plugin": "^2.7.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.5",
@@ -87,8 +85,8 @@
     "webpack-dev-server": "4.11.1"
   },
   "peerDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": ">=18.0.0",
+    "react-dom": ">=18.0.0"
   },
   "files": [
     "build"
