+++ src/table/table.tsx
@@ -23,6 +23,7 @@ import {
   isAllRowsExpanded,
 } from './utils';
 import { GenericTableProps, Row, RowAction } from './types';
+import RowGroup from './row-group';
 
 function GenericTable<T>({
   data,
@@ -242,236 +243,6 @@ function GenericTable<T>({
     ));
   };
 
-  const renderRows = (rows: Row<T>[], level = 0): ReactNode => {
-    return rows.map((row, rowIndex) => {
-      const isExpanded = expandedRows[row.id];
-      const isUpdating = updatingRowId === row.id;
-      const actionWithRenderUpdate = actions?.find(
-        (action) => action?.renderUpdateComponent
-      );
-
-      if (isUpdating && actionWithRenderUpdate) {
-        return (
-          <React.Fragment key={`${row.id}_${rowIndex}`}>
-            <TableRow
-              style={{
-                backgroundColor: "#FFFFFF",
-                cursor: onViewRow ? "pointer" : "default",
-                border: "none",
-                borderBottom: "1px solid #2F736E1F",
-              }}
-              onClick={() => onViewRow?.(row)}
-            >
-              {meta?.chartType === TABLE_TYPES.MULTI_LEVEL && (
-                <TableCell
-                  style={{
-                    width: "5px",
-                    paddingLeft: `${level * 20 + 10}px`,
-                    paddingTop: 0,
-                    paddingBottom: 0,
-                    backgroundColor: "#2F736E1F",
-                    lineHeight: "1rem",
-                    paddingRight: "0px",
-                    border: "none",
-                    borderTopLeftRadius: "8px",
-                    borderBottomLeftRadius: "8px",
-                  }}
-                ></TableCell>
-              )}
-              {columns.map((column, colIndex) => (
-                <TableCell
-                  key={String(column.key)}
-                  sx={{
-                    ...DEFAULT_TABLE_CELL_STYLES,
-                    ...tableCellStyles,
-                    border: "none",
-                    borderLeft: "1px solid #2F736E1F",
-                    background: "#2F736E1F",
-                  }}
-                >
-                  {column.render
-                    ? column.render(row[column.key as keyof Row<T>], row as T, level)
-                    : renderCellContent(row, column.key as string)}
-                </TableCell>
-              ))}
-              {isShowActionColumn && level === 0 && (
-                <TableCell
-                  key={String(`${row.id}_${rowIndex}_actions`)}
-                  sx={{
-                    ...DEFAULT_TABLE_CELL_STYLES,
-                    ...tableCellStyles,
-                    paddingY: "0.75rem",
-                    border: "none",
-                    borderLeft: "1px solid #2F736E1F",
-                    background: "#2F736E1F",
-                    borderTopRightRadius: "8px",
-                    borderBottomRightRadius: "8px",
-                  }}
-                ></TableCell>
-              )}
-            </TableRow>
-            <TableRow key={`selected_${row.id}`}>
-              <TableCell
-                colSpan={columns.length + 2}
-                style={{
-                  padding: 0,
-                  backgroundColor: "#2F736E1F",
-                  borderBottom: "none",
-                  borderRadius: "16px",
-                }}
-              >
-                {actionWithRenderUpdate.renderUpdateComponent(row, () =>
-                  setUpdatingRowId(null)
-                )}
-              </TableCell>
-            </TableRow>
-          </React.Fragment>
-        );
-      }
-
-      return (
-        <React.Fragment key={`${row.id}_${rowIndex}`}>
-          <TableRow
-            style={{
-              backgroundColor:
-              rowIndex === rows.length - 1 && showTotal
-                  ? "#2F736E"
-                  : "#F0F0F1",
-              cursor: onViewRow ? "pointer" : "default",
-            }}
-            onClick={() => onViewRow?.(row)}
-          >
-            {meta?.chartType === TABLE_TYPES.MULTI_LEVEL && (
-              <TableCell
-                style={{
-                  width: "5px",
-                  paddingLeft: `${level * 20 + 10}px`,
-                  paddingTop: 0,
-                  paddingBottom: 0,
-                  backgroundColor: "white",
-                  borderBottom: "none",
-                  lineHeight: "1rem",
-                  paddingRight: "0px",
-                }}
-              >
-                {row.children && (
-                  <button
-                    onClick={() => handleExpandRowClick(row.id)}
-                    style={{
-                      border: "none",
-                      background: "none",
-                      cursor: "pointer",
-                      padding: 0,
-                    }}
-                  >
-                    <ArrowDropDownIcon
-                      style={{
-                        color: "#162C36",
-                        rotate: isExpanded ? "0deg" : "-90deg",
-                      }}
-                    />
-                  </button>
-                )}
-                {!row.children && row.getChildren && (
-                  <button
-                    onClick={(e) =>
-                      handleExpandDynamicChildrenRowClick(rowIndex, row, e)
-                    }
-                    style={{
-                      border: "none",
-                      background: "none",
-                      cursor: "pointer",
-                      padding: 0,
-                    }}
-                  >
-                    <ArrowDropDownIcon
-                      style={{
-                        color: "#162C36",
-                        rotate:
-                        dynamicNestedRows?.index === rowIndex || isExpanded
-                            ? "0deg"
-                            : "-90deg",
-                      }}
-                    />
-                  </button>
-                )}
-              </TableCell>
-            )}
-            {columns.map((column, colIndex) => (
-              <TableCell
-                key={String(column.key)}
-                sx={{
-                  ...DEFAULT_TABLE_CELL_STYLES,
-                  ...tableCellStyles,
-                  border: "none",
-                  borderLeft:
-                  colIndex === 0 || level === depth
-                      ? "none"
-                      : "1px solid #2F736E1F",
-                  background: getRowCellBackGround(level),
-                  borderTopLeftRadius: colIndex === 0 ? "8px" : "0px",
-                  borderBottomLeftRadius: colIndex === 0 ? "8px" : "0px",
-                  borderTopRightRadius:
-                  colIndex === columns.length - 1 ? "8px" : "0px",
-                  borderBottomRightRadius:
-                  colIndex === columns.length - 1 ? "8px" : "0px",
-                }}
-              >
-                {column.render
-                  ? column.render(row[column.key as keyof Row<T>], row as T, level)
-                  : renderCellContent(row, column.key as string)}
-              </TableCell>
-            ))}
-            {isShowActionColumn && level === 0 && (
-              <TableCell className="table-cell table-cell--action">
-                {actions?.map((action, index) => {
-                  if (action?.condition && !action.condition(row.id)) {
-                    return null;
-                  }
-                  if (action.component) {
-                    return React.cloneElement(action.component, {
-                      onClick: (e: React.MouseEvent) => {
-                        if (!action.action) return;
-                        action.action(row.id, e);
-                      },
-                    });
-                  }
-                  return (
-                    <button
-                      key={String(`${row.id}_${rowIndex}_${index}_action`)}
-                      onClick={(e) => {
-                        handleRowActionClick(row.id, action);
-                        e.stopPropagation();
-                      }}
-                      className="table-button table-button--action"
-                    >
-                      {action.icon && (
-                        <img
-                          src={require(`${action.icon}`)}
-                          alt="action_icon"
-                          className="table-icon"
-                        />
-                      )}
-                      <Typography className="table-cell--action-label">
-                        {action.label}
-                      </Typography>
-                    </button>
-                  );
-                })}
-              </TableCell>
-            )}
-          </TableRow>
-          {isExpanded && row.children && renderRows(row.children, level + 1)}
-          {depth &&
-          dynamicNestedRows?.index === rowIndex &&
-          dynamicNestedRows?.dynamicRows?.length
-            ? renderDynamicRows(dynamicNestedRows.dynamicRows)
-            : null}
-        </React.Fragment>
-      );
-    });
-  };
-
   if (!data.length) {
     return (
       <div
@@ -599,7 +370,21 @@ function GenericTable<T>({
             )}
           </TableRow>
         </TableHead>
-        <TableBody>{renderRows(data)}</TableBody>
+        <TableBody>
+          <RowGroup
+            rows={data}
+            columns={columns}
+            meta={meta}
+            actions={actions}
+            onViewRow={onViewRow}
+            depth={depth}
+            showTotal={showTotal}
+            tableCellStyles={tableCellStyles}
+            rowColors={rowColors}
+            columnMetadata={columnMetadata}
+            renderCellContent={renderCellContent}
+          />
+        </TableBody>
       </Table>
     </TableContainer>
     </>
